move coefficientsToPercents to distribution slice
