+++ src/components/sections/controls/TemplateHelper.jsx
@@ -4,7 +4,7 @@ import { Typography, TextField } from '@material-ui/core';
 import { useTranslation, Trans } from 'gatsby-plugin-react-i18next';
 import { getControl } from '../../../state/controlsSlice';
 import { parseAmount } from '../../../state/optimizer/optimizerCore';
-import { Condition } from '../../../utils/gw2-data';
+import { coefficientsToPercents } from '../../../state/slices/distribution';
 
 const initial = {
   Power: 0,
@@ -19,29 +19,6 @@ const roundOne = (num) => Math.round(num * 10) / 10;
 const roundZero = (num) => Math.round(num);
 
 const indent = (str, amount) => str.replace(/^/gm, ' '.repeat(amount));
-
-// reverse legacy percent distribution conversion
-// see: https://github.com/discretize/discretize-gear-optimizer/discussions/136
-const coefficientsToPercents = (values2) => {
-  const { Power, ...rest } = values2;
-  const values1 = {};
-
-  // reverse magic numbers
-  values1.Power = (Power / 2597) * 1025;
-  for (const [key, value] of Object.entries(rest)) {
-    values1[key] = value * Condition[key].baseDamage;
-  }
-
-  // scale up/down so sum is 100
-  const sum = Object.values(values1).reduce((prev, cur) => prev + cur, 0);
-  if (sum) {
-    for (const key of Object.keys(values1)) {
-      values1[key] *= 100 / sum;
-    }
-  }
-
-  return values1;
-};
 
 // replaces "Poison" with "Poisoned" (keeping the same key order)
 const fixPoison = (input) =>
